Correction de la projection Matrice. Correction des path vers les shaders.

+++ lwjglTest/src/testArchitecture/Window.java
@@ -88,6 +88,6 @@ public class Window {
 	}
 	
 	public static double aspectRatio(){
-		return width/height;
+		return width/(double)height;
 	}
 }
